requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LLM Providers
langchain-groq>=0.2.0
langchain-openai>=0.1.0
# Примечание: langchain-google-genai, langchain-anthropic, langchain-mistralai 
# устанавливаются пользователем по желанию, если он выберет их в .env


# Vector Database and Embeddings
chromadb>=0.5.0
langchain-chroma>=0.1.4
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0

# API and Web UI
fastapi>=0.103.0
uvicorn>=0.23.2
requests>=2.31.0
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.1
pyyaml>=6.0.1
python-hcl2>=4.3.0